+++ changelog.md
@@ -1,4 +1,5 @@
 # v1.0.5
 
+ - begin implementation of libaconfig (see the readme in the libaconfig folder for more info on that)
  - allow multiple aconfig files to be selected (separated by ';')
  - more bugfixes